+++ paymentsheet/src/main/java/com/stripe/android/link/LinkControllerInteractor.kt
@@ -23,6 +23,7 @@ import com.stripe.android.link.exceptions.MissingConfigurationException
 import com.stripe.android.link.injection.LinkComponent
 import com.stripe.android.link.model.LinkAccount
 import com.stripe.android.link.model.toLoginState
+import com.stripe.android.link.theme.isDarkTheme
 import com.stripe.android.link.ui.inline.SignUpConsentAction
 import com.stripe.android.link.ui.wallet.displayName
 import com.stripe.android.link.ui.wallet.makeFallbackCardName
@@ -53,6 +54,8 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.sync.Mutex
+import kotlinx.coroutines.sync.withLock
 import org.json.JSONObject
 import javax.inject.Inject
 import javax.inject.Provider
@@ -113,11 +116,11 @@ internal class LinkControllerInteractor @Inject constructor(
         MutableSharedFlow<LinkController.PresentResult>(extraBufferCapacity = 1)
     val presentResultFlow = _presentResultFlow.asSharedFlow()
 
-    private val _presentSelectionSucceededFlow = MutableSharedFlow<Unit>(extraBufferCapacity = 1)
+    private val createPaymentMethodMutex = Mutex()
 
-    init {
+    private fun createPaymentMethodForPresentResult() {
         coroutineScope.launch {
-            _presentSelectionSucceededFlow.collect {
+            createPaymentMethodMutex.withLock {
                 val pmResult = performCreatePaymentMethod(apiKey = null)
                 updateState { it.copy(createdPaymentMethod = pmResult.getOrNull()) }
                 val presentResult = pmResult.fold(
@@ -147,15 +150,20 @@ internal class LinkControllerInteractor @Inject constructor(
 
     val selectedPaymentMethodPreview: StateFlow<LinkController.PaymentMethodPreview?> =
         _state.mapAsStateFlow { state ->
+            val isDarkTheme = state.linkConfiguration?.linkAppearance?.style.isDarkTheme(
+                isSystemDarkTheme = application.isSystemDarkTheme()
+            )
             state.selectedPaymentMethod?.details?.toPreview(
                 context = application,
                 iconLoader = cachedIconLoader,
                 reduceLinkBranding = state.linkConfiguration?.linkAppearance?.reduceLinkBranding ?: false,
+                isDarkTheme = isDarkTheme
             )
         }
 
     fun state(context: Context): StateFlow<LinkController.State> {
         return combineAsStateFlow(_internalLinkAccount, _state) { account, state ->
+            val isDarkTheme = state.linkConfiguration?.linkAppearance?.style.isDarkTheme(context.isSystemDarkTheme())
             LinkController.State(
                 elementsSessionId = state.linkConfiguration?.elementsSessionId,
                 internalLinkAccount = account,
@@ -165,6 +173,7 @@ internal class LinkControllerInteractor @Inject constructor(
                         context = context,
                         iconLoader = cachedIconLoader,
                         reduceLinkBranding = state.linkConfiguration?.linkAppearance?.reduceLinkBranding ?: false,
+                        isDarkTheme = isDarkTheme
                     ),
                 createdPaymentMethod = state.createdPaymentMethod,
             )
@@ -435,7 +444,7 @@ internal class LinkControllerInteractor @Inject constructor(
                 is LinkActivityResult.Completed -> {
                     logger.debug("$tag: present PM selected, creating payment method")
                     updateState { it.copy(selectedPaymentMethod = result.selectedPayment) }
-                    _presentSelectionSucceededFlow.tryEmit(Unit)
+                    createPaymentMethodForPresentResult()
                 }
                 is LinkActivityResult.Failed -> {
                     logger.debug("$tag: present failed")
@@ -828,7 +837,8 @@ internal fun PaymentMethodPreviewDetails.toPreview(
     iconLoader: PaymentSelection.IconLoader
 ): LinkController.PaymentMethodPreview {
     val label = context.getString(com.stripe.android.R.string.stripe_link)
-    val drawableResourceId = getIconDrawableRes(this, context.isSystemDarkTheme())
+    val isDarkTheme = context.isSystemDarkTheme()
+    val drawableResourceId = getIconDrawableRes(this, isDarkTheme)
     val sublabel = buildString {
         val name: ResolvableString
         val last4: String
@@ -865,7 +875,7 @@ internal fun PaymentMethodPreviewDetails.toPreview(
                 drawableResourceIdNight = null,
                 lightThemeIconUrl = null,
                 darkThemeIconUrl = null,
-                useDarkThemeIcon = null,
+                useDarkThemeIcon = isDarkTheme,
             )
         },
         label = label,
@@ -878,6 +888,7 @@ internal fun ConsumerPaymentDetails.PaymentDetails.toPreview(
     context: Context,
     iconLoader: PaymentSelection.IconLoader,
     reduceLinkBranding: Boolean,
+    isDarkTheme: Boolean,
 ): LinkController.PaymentMethodPreview {
     val label = context.getString(com.stripe.android.R.string.stripe_link)
     val sublabel = buildString {
@@ -889,7 +900,7 @@ internal fun ConsumerPaymentDetails.PaymentDetails.toPreview(
         append(last4)
     }
     val drawableResourceId = if (reduceLinkBranding) {
-        getIconDrawableRes(context.isSystemDarkTheme())
+        getIconDrawableRes(isDarkTheme)
     } else {
         getLinkIconArrow()
     }
@@ -913,7 +924,7 @@ internal fun ConsumerPaymentDetails.PaymentDetails.toPreview(
                 drawableResourceIdNight = null,
                 lightThemeIconUrl = null,
                 darkThemeIconUrl = null,
-                useDarkThemeIcon = null,
+                useDarkThemeIcon = isDarkTheme,
             )
         },
         label = label,